chore(constants): extract constants from PairingCard/FrontCard
